backend/requirements.txt
# Core FastAPI/Web App Requirements
fastapi
uvicorn[standard]
httpx
websockets
aiofiles
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic
pydantic-settings
pydantic[email]

---
# Database and ORM
sqlalchemy
alembic
psycopg2-binary
pgvector
lancedb
tantivy

---
# Background and Asynchronous Processing
celery
redis
flower
gevent
agno
watchdog

---
# AI and Document Processing
groq
ollama
tiktoken
cohere
langchain
langchain-community
langchain-groq
langchain-ollama
unstructured
unstructured_inference
transformers
torch
rank_bm25
striprtf
textract

---
# File and Document Handling
PyPDF2
pdfplumber
pdf2image
python-docx
openpyxl
xlrd
python-pptx
Pillow
beautifulsoup4
lxml
chardet
python-magic
tabula-py
pi_heif
six
pathlib2

---
# Data and Search
pandas
duckduckgo-search

---
# Have to downgrade to allow text textract and striprtf : C:\Projects\Allemny-Find-V2\venv\Scripts\python.exe -m pip install --upgrade pip<24.1
#pip install pypdf
# eventlet
# npm install axios
# npm install react-hot-toastls
# npm install react-markdown
# npm install remark-gfm
# npm install remark-emoji
# pip install langchain
#pip install unstructured_inference
#pip install pi_heif   
#npm install @radix-ui/react-tooltip
#pip install pdfplumber
# npm install remark-breaks
# pip install unstructured
# pip install langchain-community langchain-groq langchain-ollama
# Run this in the order for the six issue:
# 1.pip install --upgrade six
# 2.pip install --upgrade --force-reinstall --no-cache-dir python-dateutil celery
# agno uvicorn fastapi httpx
#  pip install pdf2image

## Searching
#pip install tabula
#pip install pdfplumber
#pip install pandas
#pip install transformers
#pip install torch  # for ML-based table extraction
#pip install rank_bm25  # for keyword search
#pip install tiktoken
#npm install @heroicons/react
#npm install recharts
#npm install date-fns




